:hocho: split Recipe::Entry to some module
